feat: [#65736] Export separate files for `mod11_separate_cprs`

+++ suila/suila/models.py
@@ -1458,7 +1458,7 @@ class PrismeBatch(PermissionsMixin, models.Model):
         db_index=True,
     )
 
-    prefix = models.IntegerField(
+    prefix = models.BigIntegerField(
         db_index=True,
     )
